pyproject.toml
[project]
name = "lcst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.25.15",
    "ddgs>=9.12.1",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.25",
    "langchain-openai>=1.1.12",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "langchain-deepseek>=1.0.1",
    "chromadb>=1.5.5",
]
